Use filters on pages. Fixes #20
