+++ web/components/Filters.tsx
@@ -1,27 +1,86 @@
-import React from "react";
-import { SortBy, ViewAs } from "types";
+import React, { useCallback } from "react";
+import { FilterType, SortByValue, ViewAsValue } from "types";
 
 export interface Props {
-	className?: string;
-	sortByValue?: SortBy;
+	types?: Array<FilterType>;
+	sortByValue?: SortByValue;
 	onChangeSortBy?: (event: React.ChangeEvent<HTMLSelectElement>) => void;
-	viewAsValue?: ViewAs;
+	viewAsValue?: ViewAsValue;
 	onChangeViewAs?: (event: React.ChangeEvent<HTMLSelectElement>) => void;
+	className?: string;
 }
 
 const Filters = ({
-	className,
+	types = [],
 	sortByValue,
 	onChangeSortBy,
 	viewAsValue,
 	onChangeViewAs,
+	className,
 }: Props) => {
-	if (!sortByValue && !onChangeSortBy && !viewAsValue && !onChangeViewAs) {
+	if (types.length === 0) {
 		return null;
 	}
 
-	const shouldShowSortBy = sortByValue && onChangeSortBy;
-	const shouldShowViewAs = viewAsValue && onChangeViewAs;
+	const renderSortBy = useCallback(
+		(key) => {
+			if (
+				!types.includes(FilterType.SortBy) ||
+				(!sortByValue && !onChangeSortBy)
+			) {
+				return null;
+			}
+			return (
+				<div
+					className={`flex flex-row items-center mr-0 md:mb-0 ${
+						types.length > 1 ? "mb-2 md:mr-6" : ""
+					}`}
+					key={key}
+				>
+					Sort by
+					<select
+						className="ml-2"
+						value={sortByValue}
+						onChange={onChangeSortBy}
+					>
+						<option value={SortByValue.RecentlyTagged}>Recently tagged</option>
+						<option value={SortByValue.RecentlyAdded}>Recently added</option>
+					</select>
+				</div>
+			);
+		},
+		[types, sortByValue, onChangeSortBy]
+	);
+
+	const renderViewAs = useCallback(
+		(key) => {
+			if (
+				!types.includes(FilterType.ViewAs) ||
+				(!viewAsValue && !onChangeViewAs)
+			) {
+				return null;
+			}
+			return (
+				<div
+					className={`flex flex-row items-center mr-0 md:mb-0 ${
+						types.length > 1 ? "mb-2 md:mr-6" : ""
+					}`}
+					key={key}
+				>
+					View as
+					<select
+						className="ml-2"
+						value={viewAsValue}
+						onChange={onChangeViewAs}
+					>
+						<option value={ViewAsValue.Card}>Cards</option>
+						<option value={ViewAsValue.Compact}>Compact</option>
+					</select>
+				</div>
+			);
+		},
+		[types, viewAsValue, onChangeViewAs]
+	);
 
 	return (
 		<div
@@ -29,32 +88,13 @@ const Filters = ({
 				className ?? ""
 			}`}
 		>
-			{shouldShowSortBy && (
-				<div className="flex flex-row items-center mb-2 md:mb-0 mr-0 md:mr-6">
-					Sort by
-					<select
-						className="ml-2"
-						value={sortByValue}
-						onChange={onChangeSortBy}
-					>
-						<option value={SortBy.RecentlyTagged}>Recently tagged</option>
-						<option value={SortBy.RecentlyAdded}>Recently added</option>
-					</select>
-				</div>
-			)}
-			{shouldShowViewAs && (
-				<div className="flex flex-row items-center">
-					View as
-					<select
-						className="ml-2"
-						value={viewAsValue}
-						onChange={onChangeViewAs}
-					>
-						<option value={ViewAs.Cards}>Cards</option>
-						<option value={ViewAs.Compact}>Compact</option>
-					</select>
-				</div>
-			)}
+			{types.map((type, index) => {
+				if (type === FilterType.SortBy) {
+					return renderSortBy(index);
+				} else if (type === FilterType.ViewAs) {
+					return renderViewAs(index);
+				}
+			})}
 		</div>
 	);
 };
