+++ web/pages/entities/collections/[slug].tsx
@@ -4,13 +4,14 @@ import Link from "next/link";
 import { useQuery, gql } from "@apollo/client";
 
 import { EntityFragments } from "fragments";
-import { Collection } from "types";
+import { Collection, FilterType } from "types";
 import useAudioItemsTaggedWithEntity from "hooks/useAudioItemsTaggedWithEntity";
+import useFilters from "hooks/useFilters";
 import TagService from "services/Tag";
 
 import Layout from "components/Layout";
 import LoadingBlock from "components/LoadingBlock";
-import AudioItemComponent from "components/AudioItem";
+import AudioItem from "components/AudioItem";
 import TagWithRelationshipToObject from "components/TagWithRelationshipToObject";
 import AddTagButton from "components/AddTagButton";
 import EditTagsButton from "components/EditTagsButton";
@@ -44,6 +45,10 @@ const ViewCollectionBySlug = () => {
 		{ loading: audioItemsLoading, error: audioItemsError },
 		fetchNextPageOfAudioItems,
 	] = useAudioItemsTaggedWithEntity({ entity: collection });
+
+	const { Filters, filtersProps, viewAsValue } = useFilters({
+		types: [FilterType.ViewAs],
+	});
 
 	const aboutMarkup = useMemo(
 		() => (
@@ -132,8 +137,11 @@ const ViewCollectionBySlug = () => {
 
 					{shouldShowAudioItems && (
 						<>
+							<Filters {...filtersProps} className="mb-6" />
+
 							{audioItems.map((audioItem, index) => (
-								<AudioItemComponent
+								<AudioItem
+									viewAs={viewAsValue}
 									audioItem={audioItem}
 									key={index}
 									className="mb-8"
