+++ web/pages/entities/instruments/[slug].tsx
@@ -4,13 +4,14 @@ import Link from "next/link";
 import { useQuery, gql } from "@apollo/client";
 
 import { EntityFragments } from "fragments";
-import { Instrument } from "types";
+import { Instrument, FilterType } from "types";
 import useAudioItemsTaggedWithEntity from "hooks/useAudioItemsTaggedWithEntity";
+import useFilters from "hooks/useFilters";
 import TagService from "services/Tag";
 
 import Layout from "components/Layout";
 import LoadingBlock from "components/LoadingBlock";
-import AudioItemComponent from "components/AudioItem";
+import AudioItem from "components/AudioItem";
 import AddTagButton from "components/AddTagButton";
 import EditTagsButton from "components/EditTagsButton";
 import TagWithRelationshipToObject from "components/TagWithRelationshipToObject";
@@ -44,6 +45,10 @@ const ViewInstrumentBySlug = () => {
 		{ loading: audioItemsLoading, error: audioItemsError },
 		fetchNextPageOfAudioItems,
 	] = useAudioItemsTaggedWithEntity({ entity: instrument });
+
+	const { Filters, filtersProps, viewAsValue } = useFilters({
+		types: [FilterType.ViewAs],
+	});
 
 	const aboutMarkup = useMemo(
 		() => (
@@ -119,8 +124,11 @@ const ViewInstrumentBySlug = () => {
 
 					{shouldShowAudioItems && (
 						<>
+							<Filters {...filtersProps} className="mb-6" />
+
 							{audioItems.map((audioItem, index) => (
-								<AudioItemComponent
+								<AudioItem
+									viewAs={viewAsValue}
 									audioItem={audioItem}
 									key={index}
 									className="mb-8"
