+++ web/pages/entities/tunes/[slug].tsx
@@ -3,13 +3,14 @@ import { useRouter } from "next/router";
 import { useQuery, gql } from "@apollo/client";
 
 import { EntityFragments } from "fragments";
-import { Tune } from "types";
+import { Tune, FilterType } from "types";
 import useAudioItemsTaggedWithEntity from "hooks/useAudioItemsTaggedWithEntity";
+import useFilters from "hooks/useFilters";
 import TagService from "services/Tag";
 
 import Layout from "components/Layout";
 import LoadingBlock from "components/LoadingBlock";
-import AudioItemComponent from "components/AudioItem";
+import AudioItem from "components/AudioItem";
 import TagWithRelationshipToObject from "components/TagWithRelationshipToObject";
 import AddTagButton from "components/AddTagButton";
 import EditTagsButton from "components/EditTagsButton";
@@ -45,6 +46,10 @@ const ViewTuneBySlug = () => {
 		{ loading: audioItemsLoading, error: audioItemsError },
 		fetchNextPageOfAudioItems,
 	] = useAudioItemsTaggedWithEntity({ entity: tuneData?.tune });
+
+	const { Filters, filtersProps, viewAsValue } = useFilters({
+		types: [FilterType.ViewAs],
+	});
 
 	const aboutMarkup = useMemo(
 		() => (
@@ -155,8 +160,11 @@ const ViewTuneBySlug = () => {
 
 					{shouldShowAudioItems && (
 						<>
+							<Filters {...filtersProps} className="mb-6" />
+
 							{audioItems.map((audioItem, index) => (
-								<AudioItemComponent
+								<AudioItem
+									viewAs={viewAsValue}
 									audioItem={audioItem}
 									key={index}
 									className="mb-8"
